+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.2.1 - 2024-07-12
+### Added
+- Added config option `paysera_logging_extra.sentry.minimum_log_level` to set from which level logs will be sent to Sentry
+- Added config option `paysera_logging_extra.monolog.minimum_introspection_level` to set from which level logs will be introspected
+
 ## 2.2.0
 ### Added
 - Added Symfony ^6 support.